requirements.txt
aiofiles==24.1.0 ; python_version == "3.12"
aiogram==3.23.0 ; python_version == "3.12"
aiohappyeyeballs==2.6.1 ; python_version == "3.12"
aiohttp==3.13.2 ; python_version == "3.12"
aiosignal==1.4.0 ; python_version == "3.12"
alembic==1.18.1 ; python_version == "3.12"
annotated-doc==0.0.4 ; python_version == "3.12"
annotated-types==0.7.0 ; python_version == "3.12"
anyio==4.12.0 ; python_version == "3.12"
attrs==25.4.0 ; python_version == "3.12"
beautifulsoup4==4.14.3 ; python_version == "3.12"
bs4==0.0.2 ; python_version == "3.12"
cachetools==6.2.4 ; python_version == "3.12"
certifi==2025.11.12 ; python_version == "3.12"
charset-normalizer==3.4.4 ; python_version == "3.12"
click==8.3.1 ; python_version == "3.12"
colorama==0.4.6 ; (platform_system == "Windows" or sys_platform == "win32") and python_version == "3.12"
distro==1.9.0 ; python_version == "3.12"
dnspython==2.8.0 ; python_version == "3.12"
email-validator==2.3.0 ; python_version == "3.12"
fastapi-cli==0.0.16 ; python_version == "3.12"
fastapi-cloud-cli==0.5.2 ; python_version == "3.12"
fastapi==0.123.4 ; python_version == "3.12"
fastar==0.8.0 ; python_version == "3.12"
filelock==3.20.1 ; python_version == "3.12"
frozenlist==1.8.0 ; python_version == "3.12"
fsspec==2025.12.0 ; python_version == "3.12"
google-ai-generativelanguage==0.6.15 ; python_version == "3.12"
google-api-core==2.28.1 ; python_version == "3.12"
google-api-python-client==2.187.0 ; python_version == "3.12"
google-auth-httplib2==0.3.0 ; python_version == "3.12"
google-auth==2.45.0 ; python_version == "3.12"
google-genai==1.56.0 ; python_version == "3.12"
google-generativeai==0.8.6 ; python_version == "3.12"
googleapis-common-protos==1.72.0 ; python_version == "3.12"
greenlet==3.3.0 ; python_version == "3.12"
grpcio-status==1.71.2 ; python_version == "3.12"
grpcio==1.76.0 ; python_version == "3.12"
h11==0.16.0 ; python_version == "3.12"
httpcore==1.0.9 ; python_version == "3.12"
httplib2==0.31.0 ; python_version == "3.12"
httptools==0.7.1 ; python_version == "3.12"
httpx==0.28.1 ; python_version == "3.12"
idna==3.11 ; python_version == "3.12"
jinja2==3.1.6 ; python_version == "3.12"
jiter==0.12.0 ; python_version == "3.12"
joblib==1.5.3 ; python_version == "3.12"
llvmlite==0.46.0 ; python_version == "3.12"
lxml==6.0.2 ; python_version == "3.12"
magic-filter==1.0.12 ; python_version == "3.12"
mako==1.3.10 ; python_version == "3.12"
markdown-it-py==4.0.0 ; python_version == "3.12"
markupsafe==3.0.3 ; python_version == "3.12"
mdurl==0.1.2 ; python_version == "3.12"
more-itertools==10.8.0 ; python_version == "3.12"
mpmath==1.3.0 ; python_version == "3.12"
multidict==6.7.0 ; python_version == "3.12"
networkx==3.6.1 ; python_version == "3.12"
numba==0.63.1 ; python_version == "3.12"
numpy==2.3.5 ; python_version == "3.12"
nvidia-cublas-cu12==12.8.4.1 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version == "3.12"
nvidia-cuda-cupti-cu12==12.8.90 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version == "3.12"
nvidia-cuda-nvrtc-cu12==12.8.93 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version == "3.12"
nvidia-cuda-runtime-cu12==12.8.90 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version == "3.12"
nvidia-cudnn-cu12==9.10.2.21 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version == "3.12"
nvidia-cufft-cu12==11.3.3.83 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version == "3.12"
nvidia-cufile-cu12==1.13.1.3 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version == "3.12"
nvidia-curand-cu12==10.3.9.90 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version == "3.12"
nvidia-cusolver-cu12==11.7.3.90 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version == "3.12"
nvidia-cusparse-cu12==12.5.8.93 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version == "3.12"
nvidia-cusparselt-cu12==0.7.1 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version == "3.12"
nvidia-nccl-cu12==2.27.5 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version == "3.12"
nvidia-nvjitlink-cu12==12.8.93 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version == "3.12"
nvidia-nvshmem-cu12==3.3.20 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version == "3.12"
nvidia-nvtx-cu12==12.8.90 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version == "3.12"
openai-whisper==20250625 ; python_version == "3.12"
openai==2.13.0 ; python_version == "3.12"
playwright==1.58.0 ; python_version == "3.12"
propcache==0.4.1 ; python_version == "3.12"
proto-plus==1.27.0 ; python_version == "3.12"
protobuf==5.29.5 ; python_version == "3.12"
psycopg2-binary==2.9.11 ; python_version == "3.12"
psycopg2==2.9.11 ; python_version == "3.12"
pyasn1-modules==0.4.2 ; python_version == "3.12"
pyasn1==0.6.1 ; python_version == "3.12"
pydantic-core==2.33.2 ; python_version == "3.12"
pydantic==2.11.10 ; python_version == "3.12"
pyee==13.0.0 ; python_version == "3.12"
pygments==2.19.2 ; python_version == "3.12"
pyparsing==3.2.5 ; python_version == "3.12"
python-dotenv==1.2.1 ; python_version == "3.12"
python-multipart==0.0.20 ; python_version == "3.12"
pyyaml==6.0.3 ; python_version == "3.12"
regex==2025.11.3 ; python_version == "3.12"
requests==2.32.5 ; python_version == "3.12"
rich-toolkit==0.17.0 ; python_version == "3.12"
rich==14.2.0 ; python_version == "3.12"
rignore==0.7.6 ; python_version == "3.12"
rsa==4.9.1 ; python_version == "3.12"
scikit-learn==1.8.0 ; python_version == "3.12"
scipy==1.17.0 ; python_version == "3.12"
sentry-sdk==2.46.0 ; python_version == "3.12"
setuptools==80.9.0 ; python_version == "3.12"
shellingham==1.5.4 ; python_version == "3.12"
six==1.17.0 ; python_version == "3.12"
sniffio==1.3.1 ; python_version == "3.12"
soupsieve==2.8.3 ; python_version == "3.12"
sqlalchemy==2.0.45 ; python_version == "3.12"
starlette==0.50.0 ; python_version == "3.12"
sympy==1.14.0 ; python_version == "3.12"
tenacity==9.1.2 ; python_version == "3.12"
threadpoolctl==3.6.0 ; python_version == "3.12"
tiktoken==0.12.0 ; python_version == "3.12"
torch==2.9.1 ; python_version == "3.12"
tqdm==4.67.1 ; python_version == "3.12"
triton==3.5.1 ; (platform_machine == "x86_64" or sys_platform == "linux2") and python_version == "3.12" and (platform_system == "Linux" or sys_platform == "linux" or sys_platform == "linux2")
typer==0.20.0 ; python_version == "3.12"
typing-extensions==4.15.0 ; python_version == "3.12"
typing-inspection==0.4.2 ; python_version == "3.12"
uritemplate==4.2.0 ; python_version == "3.12"
urllib3==2.5.0 ; python_version == "3.12"
uvicorn==0.38.0 ; python_version == "3.12"
uvloop==0.22.1 ; sys_platform != "win32" and sys_platform != "cygwin" and platform_python_implementation != "PyPy" and python_version == "3.12"
watchfiles==1.1.1 ; python_version == "3.12"
websockets==15.0.1 ; python_version == "3.12"
whisper==1.1.10 ; python_version == "3.12"
yarl==1.22.0 ; python_version == "3.12"
